% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivhetLc.R
\name{ivhetLc}
\alias{ivhetLc}
\title{Estimation of instrumental variables models with latent classes}
\usage{
ivhetLc(
  formula,
  data,
  subset,
  na.action,
  method = "bfgs",
  model = c("lciv", "lc"),
  Q = 2,
  gradient = TRUE,
  print.init = FALSE,
  messages = TRUE,
  start = NULL,
  pert.init = 0.05,
  ...
)
}
\description{
Estimation of instrumental variables models with latent classes
}
\author{
Mauricio Sarrias
}
